python/src/users/requirements.txt
click==8.1.3
dnspython==2.2.1
Flask==2.2.2
Flask-PyMongo==2.3.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pika==1.3.1
PyJWT==2.6.0
pymongo==4.3.3
Werkzeug==2.2.2
